requirements.txt
coreapi
django
django-environ
djangorestframework
mysqlclient
gunicorn
django-prometheus